requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
axial_positional_embedding==0.3.12
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.8.1
fastjsonschema==2.21.1
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
hyper-connections==0.1.15
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lightning-utilities==0.14.0
littleutils==0.2.4
local-attention==1.11.1
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
memory-profiler==0.61.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.3
ogb==1.3.6
outdated==0.2.2
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
performer-pytorch==1.1.4
pickleshare==0.7.5
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
PyGCL==0.1.2
Pygments==2.19.1
pygrok==1.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
rdkit==2024.9.5
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.22.0
setproctitle==1.3.5
setuptools==76.0.0
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch-geometric==2.6.1
torch_scatter==2.1.2
torch_sparse==0.6.18
torchmetrics==1.6.2
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wandb==0.19.8
wcwidth==0.2.13
yacs==0.1.8
yarl==1.18.3
